lib/workload/stateless/stacks/pg-dd/pyproject.toml
[tool.poetry]
name = "postgres-data-dump"
version = "0.1.0"
description = "A serivce to dump postgres databases to S3."
authors = ["Marko malenic <mmalenic1@gmail.com>"]
readme = "README.md"
packages = [{ include = "pg_dd" }]

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1"
psycopg = { version = "^3", extras = ["binary"] }
python-dotenv = "^1"
libumccr = "^0.4"
mypy-boto3-s3 = "^1"
click = "^8"

[tool.poetry.group.dev.dependencies]
mypy = "^1"
ruff = "^0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli = "pg_dd.cli:cli"
